+++ pkgs/intl4x/lib/src/data.dart
@@ -1,0 +1,15 @@
+import 'dart:typed_data';
+
+abstract final class Data {}
+
+final class JsonData extends Data {
+  final String value;
+
+  JsonData(this.value);
+}
+
+final class BlobData extends Data {
+  final Uint8List value;
+
+  BlobData(this.value);
+}
